Refactor viewer platform surfaces
