+++ scripts/test-web-preview.sh
@@ -180,9 +180,10 @@ const xyzFast = fs.readFileSync(xyzFastPath, 'utf8');
 const configSource = fs.readFileSync(configPath, 'utf8');
 const dataSource = fs.readFileSync(dataPath, 'utf8');
 const contentView = fs.readFileSync('App/ContentView.swift', 'utf8');
-const appDelegate = fs.readFileSync('App/BurreteApp.swift', 'utf8');
-const appViewer = fs.readFileSync('App/MoleculeViewerWindowController.swift', 'utf8');
-const quickLookViewer = fs.readFileSync('PreviewExtension/PreviewViewController.swift', 'utf8');
+const appDelegate = fs.readFileSync('App/Platform/PlatformBridge.swift', 'utf8');
+const appViewer = fs.readFileSync('App/Platform/MoleculeViewerWindowController.swift', 'utf8');
+const appViewerScene = fs.readFileSync('App/MoleculeViewerScene.swift', 'utf8');
+const quickLookViewer = fs.readFileSync('PreviewExtension/Platform/PreviewViewController.swift', 'utf8');
 const quickLookInfo = fs.readFileSync('PreviewExtension/Info.plist', 'utf8');
 const appInfo = fs.readFileSync('App/Info.plist', 'utf8');
 const gridBuilder = fs.readFileSync('PreviewExtension/MoleculeGridPreview.swift', 'utf8');
@@ -261,7 +262,7 @@ assert(!appViewer.includes('exitFullScreenMode'), 'standalone app viewer must no
 assert(!appDelegate.includes('.enterFullScreen()'), 'opening files in the app must not force native fullscreen');
 assert(appViewer.includes('func reloadDisplayPreferences()'), 'app viewer must expose a display-preference reload hook');
 assert(appViewer.includes('applyWindowDisplayPreferences'), 'app viewer must update native window appearance and material preferences');
-assert(appViewer.includes('NSVisualEffectView'), 'app viewer transparency must use native macOS material instead of a fully clear custom window');
+assert(appViewerScene.includes('.regularMaterial'), 'app viewer transparency must use native SwiftUI material instead of a fully clear custom window');
 assert(appViewer.includes('viewerWindowOpacity'), 'app viewer must expose configurable window material opacity');
 assert(appViewer.includes('viewerOverlayOpacity'), 'app viewer must expose configurable overlay readability');
 assert(appViewer.includes('defaultViewerPageZoom: CGFloat = 1.0'), 'app viewer must keep WKWebView page zoom at 1.0 so Mol* mouse picking stays aligned');
